+++ Changelog.md
@@ -21,6 +21,8 @@ Breaking Changes:
   option. (Phil Pirozhkov, #2864)
 * Remove `it_should_behave_like` nested shared group method and
   `alias_it_should_behave_like_to` configuration option. (Phil Pirozhkov, #2864)
+* Remove deprecated `treat_symbols_as_metadata_keys_with_true_values` configuration
+  option. (Phil Pirozhkov, #2864)
 
 Enhancements: